+++ src/main/java/com/backend/komeet/chat/repositories/ChatRoomQRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.backend.komeet.chat.repositories;
 
-import com.backend.komeet.domain.entities.QChat;
-import com.backend.komeet.domain.entities.QChatRoom;
+import com.backend.komeet.chat.model.entities.QChat;
+import com.backend.komeet.chat.model.entities.QChatRoom;
 import com.backend.komeet.user.model.entities.User;
 import com.backend.komeet.base.model.dtos.ChatRoomDto;
 import com.querydsl.core.types.OrderSpecifier;
